See merge request s07-final/S07P31A307!266

+++ backend/src/main/java/com/ssapin/backend/api/domain/dto/response/PlaceInfoResponse.java
@@ -10,8 +10,8 @@ public class PlaceInfoResponse {
     private long placeId;
     private long itemId;
     private String title;
-    private float lat;
-    private float lng;
+    private double lat;
+    private double lng;
     private String address;
     private boolean isBookmark;
